namespacefs: Introduce 'namespacefs'
Here we introduce a simple, read-only pseudo file system that aims
to provide direct mechanism for examining the existing hierarchy of
namespaces on the system. When fully functional, 'namespacefs' will
allow the user to see all namespaces that are active on the system
and to easily retrieve the PIDs of all tasks enclosed in each
individual namespace. Such possibility can be very useful in the
context of containerized workloads. It will provide universal
methods to detect, examine and monitor all sorts of containers
running on the system, without relaying on any specific user-space
software.

Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>